refactor(agent): update test imports to reflect merged sub-domains
