+++ packages/agent/test/agent/discovery.test.ts
@@ -1,7 +1,6 @@
 import { describe, it, expect, beforeEach } from "bun:test";
 import { join } from "node:path";
-import { AgentDiscovery } from "../../src/agent/discovery/discovery";
-import { parseFrontmatter } from "../../src/agent/discovery/frontmatter";
+import { AgentDiscovery, parseFrontmatter } from "../../src/agent/discovery";
 import { BuiltinAgentRegistry } from "../../src/agent/registry/registry";
 
 const FIXTURES_DIR = join(import.meta.dir, "../fixtures/agents");
