+++ packages/agent/test/agent/file-registry-storage.test.ts
@@ -3,7 +3,7 @@ import { mkdtempSync, rmSync, existsSync } from "node:fs";
 import { join } from "node:path";
 import { tmpdir } from "node:os";
 import { FileAgentRegistryStore } from "../../src/agent/registry/file-registry-storage";
-import type { AgentProfile } from "../../src/agent/definition/profile";
+import type { AgentProfile } from "../../src/agent/definitions";
 import { AgentRegistry } from "../../src/agent/registry/profile-store";
 
 function makeProfile(id: string, name?: string): AgentProfile {
